% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{buildDataFrame}
\alias{buildDataFrame}
\title{Build data frame with points that can be plotted over a sine wave}
\usage{
buildDataFrame(dat)
}
\arguments{
\item{dat}{abundance data frame}
}
\value{
a data frame
}
\description{
Build data frame for downstream plotting
}
\examples{
buildDataFrame(abundances)
}
