% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{buildPointsPerTaxon}
\alias{buildPointsPerTaxon}
\title{Build points for each feature}
\usage{
buildPointsPerTaxon(taxon, n)
}
\arguments{
\item{taxon}{taxon name}

\item{n}{number of repititions as calculated using pointNumber()}
}
\value{
a character vector
}
\description{
Generate a list of points for a given feature
}
\examples{
buildPointsPerTaxon("E.coli", 100)
}
