% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{checkValue}
\alias{checkValue}
\title{Check the type of abundance table}
\usage{
checkValue(x)
}
\arguments{
\item{x}{data frame of abundances}
}
\value{
a string (fraction, percentage or count)
}
\description{
Check the nature of abundance (count, fraction or percentage)
}
\examples{
checkValue(counts.table)
}
