% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genus_abundances}
\alias{genus_abundances}
\title{Dataset
Data set of microbial abundances for two samples}
\format{data frame}
\usage{
data(genus_abundances)
}
\description{
Dataset
Data set of microbial abundances for two samples
}
\keyword{datasets}
