% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{microwave}
\alias{microwave}
\title{Plot a wave based on microbial profiles}
\usage{
microwave(abundances, with_image = TRUE, background = "white",
  text = "")
}
\arguments{
\item{abundances}{a data frame object}

\item{with_image}{do you want to include the epithelium? Deafults to TRUE}

\item{background}{colour of background in the plots (colours supported in the
generic R colour palettes)}

\item{text}{include text above the plot. Defaults to no text.}
}
\value{
.png files with plots per sample (columns in abundance data frame)
}
\description{
Plot a wave representation of bacteria based on relative abundance data. Microwave
takes as input a data frame of microbial abundances and generates a set of points for
each taxon that are randomly asigned a position in a wave with an arbitrary colour and
size. This produces an image with an aesthetically pleasing view of an individual's
microbiome profile with the intended use for participant engagement. The aim is that
participants can receive an abstract view of their own microbiome. The wave can be produced
with or without a corresponding epithelial layer - maybe of use to folk that do not work
on gut samples but would like a visual representation of individual microbiomes.
}
\examples{
microwave(genus_abundances, with_image=FALSE)
}
