% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{plotWave}
\alias{plotWave}
\title{Plot wave with or without epithelium using ggplot2}
\usage{
plotWave(dat, cols, g = "none", with_image = TRUE,
  background = "white", text = "")
}
\arguments{
\item{dat}{data frame created with buildDataFrame}

\item{cols}{colours used for plotting (setColours())}

\item{g}{image file (epithelium)}

\item{with_image}{do you want to include the epithelium? Deafults to TRUE}

\item{background}{colour of background in the plots}

\item{text}{include text above the plot}
}
\value{
a ggplot object (grob)
}
\description{
Plot wave for each sample
}
\examples{
data(genus_abundances)
plotWave(genus_abundances, with_image=FALSE)
}
