% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{pointNumber}
\alias{pointNumber}
\title{Number of points to generate for a given feature}
\usage{
pointNumber(value)
}
\arguments{
\item{value}{abundance in fraction from which to calculate the number of points}
}
\value{
int
}
\description{
Generate number of points to generate for a feature
}
\examples{
pointNumber(0.5)
}
