% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{readAbundanceFile}
\alias{readAbundanceFile}
\title{Read in an abundance file}
\usage{
readAbundanceFile(abundanceFile)
}
\arguments{
\item{abundanceFile}{.tsv file with first column as feature. Columns are samples}
}
\value{
a data frame
}
\description{
Reads in data file
}
\examples{
readAbundanceFile("abundances.tsv")
}
