% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{relab}
\alias{relab}
\title{Relative abundance calculation}
\usage{
relab(counts)
}
\arguments{
\item{counts}{counts table (rownames are features)}
}
\value{
a data frame
}
\description{
Calculate the relative abundance (fraction) from a table of counts
}
\examples{
relab(counts.table)
}
