% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{setColours}
\alias{setColours}
\title{Set the colours for ggplot}
\usage{
setColours(dat)
}
\arguments{
\item{dat}{abundance data frame containing taxa for colouring}
}
\value{
a character vector
}
\description{
Use rainbow() for generating colours to use for points
}
\examples{
abundances <- readAbundanceFile("test.tsv")
colours <- setColours(abundances)
}
