% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micRowave.R
\name{subsetDataFrame}
\alias{subsetDataFrame}
\title{Subset a data frame by column number}
\usage{
subsetDataFrame(dat, columnNumber = 1)
}
\arguments{
\item{dat}{abundance data frame}

\item{columnNumber}{column number to take}
}
\value{
a data frame
}
\description{
Subset a data frame for plotting each sample
}
\examples{
subsetDataFrame(abundances, columnNumber=1)
}
